chatbots/polybot/{{cookiecutter.folder_name}}/app/requirements.txt
asyncio
chromadb
fastapi
httpx
langchain
openai==0.28
sseclient-py
uvicorn
tiktoken
unstructured[local-inference]
websockets